Make if-modified-since check more robust
Normally the value for if-modified-since which is send from the client is the
same timestamp. There are clients however (tt-rss) which send newer timestamps.
In this case the server should also return a 304 not modified. Older apache
servers choke on this and return a 500 "Internal Error" newer servers ignore
this but caching is broken.